# swagger.yaml
openapi: 3.0.3
info:
  title: REST API - Gestion de produits, catégories et entreprises
  description: |
    API REST complète pour la gestion d'un système e-commerce avec :
    - Gestion des produits (simples et variables)
    - Gestion des catégories et sous-catégories
    - Gestion des entreprises
    - Système de commandes
    - Authentification et autorisation basée sur les rôles
    - Notifications système
    - Upload et optimisation d'images via Cloudinary
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: ISC
servers:
  - url: http://localhost:5000
    description: Serveur de développement
  - url: https://api.yourdomain.com
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /api/auth/login
  schemas:
...
    Category:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          example: "Vêtements"
        description:
          type: string
        image:
          type: string
          description: URL de l'image
        createdAt:
          type: string
          format: date-time
          
    # Schémas des entreprises
    Company:
      type: object
      properties:
        _id:
          type: string
        nom:
          type: string
          example: "Fashion Corp"
        logo:
          type: string
          description: URL du logo
        createdAt:
          type: string
          format: date-time
          
    # Schémas d'erreur
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
  responses:
    UnauthorizedError:
      description: Token d'authentification manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Lister les catégories
      description: Récupère toutes les catégories
      responses:
        '200':
          description: Liste des catégories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
                  
    post:
      tags:
        - Categories
      summary: Créer une catégorie
      description: Crée une nouvelle catégorie (permission requise)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Catégorie créée avec succès
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Récupérer une catégorie
      description: Récupère une catégorie spécifique par son ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la catégorie
      responses:
        '200':
          description: Catégorie trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Routes des entreprises
  /api/companies:
    get:
      tags:
        - Companies
      summary: Lister les entreprises
      description: Récupère toutes les entreprises
      responses:
        '200':
          description: Liste des entreprises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
                  
    post:
      tags:
        - Companies
      summary: Créer une entreprise
      description: Crée une nouvelle entreprise (permission requise)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nom
              properties:
                nom:
                  type: string
      responses:
        '201':
          description: Entreprise créée avec succès
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/companies/{id}:
    get:
      tags:
        - Companies
      summary: Récupérer une entreprise
      description: Récupère une entreprise spécifique par son ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de l'entreprise
      responses:
        '200':
          description: Entreprise trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '404':
          $ref: '#/components/responses/NotFoundError'